+++ packages/cli/src/core/index.ts
@@ -1,9 +1,13 @@
 import packageJson from "../../package.json";
 import { Command } from "commander";
 import chalk from "chalk";
-import generateDandoriTasks, { DandoriTask } from "@dandori/core";
+import generateDandoriTasks, {
+  ChatGPTFunctionCallModel,
+  DandoriTask,
+  OptionalTaskPropsOption,
+} from "@dandori/core";
 import { readFile } from "fs/promises";
-import { generateDandoriFilePath } from "@dandori/libs";
+import { loadFile } from "@dandori/libs";
 
 export default class DandoriCoreCli {
   private inputFile: string = "";
@@ -24,7 +28,15 @@ export default class DandoriCoreCli {
       .version(packageJson.version)
       .argument("<input-file>")
       .usage(`${chalk.green("<input-file>")} [options]`)
-      .option("-e, --env-file <env-file>")
+      .option("-e, --env-file <env-file>", "env file path")
+      .option(
+        "-m, --model <model>",
+        "Chat GPT model which supports function_calling",
+      )
+      .option(
+        "-o, --optional-task-props <optional-task-props>",
+        "optional output task props which delimiter is a comma",
+      )
       .action((iFile) => {
         this.inputFile = iFile;
       });
@@ -32,11 +44,19 @@ export default class DandoriCoreCli {
 
   protected async generateDandoriTasks(): Promise<DandoriTask[]> {
     this.program.parse(process.argv);
-    const opts = this.program.opts();
-    const source = await readFile(generateDandoriFilePath(this.inputFile));
-    const { envFile } = opts;
+    const source = await readFile(loadFile(this.inputFile));
+    const { envFile, optionalTaskProps, model } = this.program.opts<{
+      envFile?: string;
+      optionalTaskProps?: string;
+      model?: ChatGPTFunctionCallModel;
+    }>();
+    // TODO: Error Handling of invalid options
     return generateDandoriTasks(source.toString(), {
       envFilePath: envFile,
+      chatGPTModel: model,
+      optionalTaskProps: optionalTaskProps?.split(
+        ",",
+      ) as OptionalTaskPropsOption,
     });
   }
 }